applications/backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
boto3==1.28.85
gunicorn==21.2.0